+++ src/main/java/com/groom/tennis_match/auth/dto/request/AdminAccountRegisterDTO.java
@@ -1,4 +1,4 @@
-package com.groom.tennis_match.auth.dto;
+package com.groom.tennis_match.auth.dto.request;
 
 import com.groom.tennis_match.auth.AdminRole;
 import lombok.*;
